% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/get_packages_by_view.R
\name{get_packages_by_view}
\alias{get_packages_by_view}
\alias{get_packages_by_views}
\title{Get Packages by biocViews}
\usage{
get_packages_by_view(view, pkg_list)

get_packages_by_views(views, pkg_list)
}
\arguments{
\item{view}{A single biocView term.}

\item{pkg_list}{Value of a call to \code{get_all_biocpkglist()}.}

\item{views}{A character vector of biocView terms.}
}
\value{
A tibble of two columns giving
the name of the package
and the Bioconductor repository it comes from.
}
\description{
\code{get_packages_by_view()} returns package names associated with a single biocViews term.

\code{get_packages_by_views()} returns packages names at the intersection of multiple biocViews terms.
}
\examples{
biocpkglist <- get_all_biocpkglist(verbose = FALSE)

get_packages_by_view(
  view = "Spatial",
  pkg_list = biocpkglist
)
get_packages_by_views(
  views = c("Spatial", "SingleCell"),
  pkg_list = biocpkglist
)
}
