% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/wordcloud.R
\name{get_view_cooccurrence_counts}
\alias{get_view_cooccurrence_counts}
\title{Get biocViews Co-Occurrence Counts}
\usage{
get_view_cooccurrence_counts(view, pkg_list = NULL, keep_self = FALSE)
}
\arguments{
\item{view}{A biocViews term.}

\item{pkg_list}{Value of a call to \code{biocPkgList()}.
If \code{NULL} (default), will call \code{biocPkgList()} internally.
See Details.}

\item{keep_self}{Include \code{view} itself in the output.}
}
\value{
Named integer vector of co-occurrence counts for each biocViews term.
}
\description{
Get biocViews Co-Occurrence Counts
}
\details{
Calling \code{BiocPkgTools::biocPkgList()} and passing the result to
\code{get_packages_by_view()} or \code{get_packages_by_views()} is more efficient
if you are making multiple calls.
See vignette 'Optimisations' for a more comprehensive discussion and demonstration.
}
\examples{
out <- get_view_cooccurrence_counts("Spatial")
head(sort(out, decreasing = TRUE))
}
