% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/get_view_membership_matrix.R
\name{get_view_membership_matrix}
\alias{get_view_membership_matrix}
\title{Package-by-biocView Membership Matrix}
\usage{
get_view_membership_matrix(pkg_list)
}
\arguments{
\item{pkg_list}{Value of a call to \code{get_all_biocpkglist()}.}
}
\value{
A logical matrix indicating packages associated with each biocViews term (or one of its child terms).
}
\description{
Package-by-biocView Membership Matrix
}
\examples{
out <- get_view_membership_matrix()
out[1:5, 1:5]
}
