% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/example.R
\name{my_bioc_function}
\alias{my_bioc_function}
\title{Example Function}
\usage{
my_bioc_function(x, y)
}
\arguments{
\item{x}{One number.}

\item{y}{Another number.}
}
\value{
The sum of \code{x}and \code{y}
}
\description{
Messages the user and returns the sum of \code{x}and \code{y}.
}
\examples{
my_bioc_function(2, 3)
}
