% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/count_packages_over_time.R
\name{count_packages_over_time}
\alias{count_packages_over_time}
\title{Track a Set of Packages Over Time}
\usage{
count_packages_over_time(packages, pkg_list = NULL, pkg_years = NULL)
}
\arguments{
\item{packages}{Character vector of package names.}

\item{pkg_list}{Value of a call to \code{biocPkgList()}.
If \code{NULL} (default), will call \code{biocPkgList()} internally.
See Details.}

\item{pkg_years}{Value of a call to \code{getPkgYearsInBioc()}.
If \code{NULL} (default), will call \code{getPkgYearsInBioc()} internally.}
}
\value{
A tibble of two columns: date and count.
\code{date} is a sequence of dates spaced by six months from 2006 to the current year.
\code{count} is the number of packages in Bioconductor on that date.
}
\description{
Given a set of packages, this function counts how many of them were present in Bioconductor over time.
}
\examples{
biocpkglist <- get_all_biocpkglist(verbose = FALSE)
count_packages_over_time(biocpkglist$Package)
}
