% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/count_packages_over_time.R
\name{count_packages_over_time}
\alias{count_packages_over_time}
\title{Track a Set of Packages Over Time}
\usage{
count_packages_over_time(packages, pkg_list, pkg_years)
}
\arguments{
\item{packages}{Character vector of package names.}

\item{pkg_list}{Value of a call to \code{get_all_biocpkglist()}.}

\item{pkg_years}{Value of a call to \code{getPkgYearsInBioc()}.}
}
\value{
A tibble of two columns: date and count.
\code{date} is a sequence of dates spaced by six months from 2006 to the current year.
\code{count} is the number of packages in Bioconductor on that date.
}
\description{
Given a set of packages, this function counts how many of them were present in Bioconductor over time.
}
\examples{
suppressPackageStartupMessages(library(BiocPkgTools))

bioc_pkg_list <- get_all_biocpkglist(verbose = FALSE)
bioc_years <- BiocPkgTools::getPkgYearsInBioc()

count_packages_over_time(
  packages = bioc_pkg_list$Package,
  pkg_list = bioc_pkg_list,
  pkg_years = bioc_years
)
}
