% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/get_all_biocpkglist.R
\name{get_all_biocpkglist}
\alias{get_all_biocpkglist}
\title{Combined Package Listing from all Bioconductor Repositories}
\usage{
get_all_biocpkglist()
}
\value{
A data frame with one row per package and columns for package metadata,
including the repository it belongs to.
}
\description{
This function is a convenience wrapper that invokes \code{\link[BiocPkgTools:biocPkgList]{BiocPkgTools::biocPkgList()}} for each of the known Bioconductor repositories ("BioCsoft", "BioCexp", "BioCworkflows", and "BioCann") and combines the results into a single data frame.
}
\examples{
biocpkglist <- get_all_biocpkglist()
}
