% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/get_packages_by_author.R
\name{get_packages_by_author}
\alias{get_packages_by_author}
\title{Get Packages by Author}
\usage{
get_packages_by_author(author, pkg_list)
}
\arguments{
\item{author}{A single author (and their aliases).}

\item{pkg_list}{Value of a call to \code{get_all_biocpkglist()}.}
}
\value{
A named list with two elements: \code{Maintainer} and \code{Author}.
}
\description{
\code{get_packages_by_author()} returns package names associated with a single author.
}
\examples{
author <- list(
  "Kevin Rue" = c(
    "Kevin Rue",
    "Kevin Rue-Albrecht"
  )
)
get_packages_by_author(author)
}
