% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/example.R
\name{get_packages_by_view}
\alias{get_packages_by_view}
\title{Get Packages by biocViews}
\usage{
get_packages_by_view(view)
}
\arguments{
\item{view}{One biocView term.}
}
\value{
Character vector of package names.
}
\description{
Get Packages by biocViews
}
\examples{
get_packages_by_view("Sequencing")
}
