% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/example.R
\name{get_packages_by_view}
\alias{get_packages_by_view}
\title{Get Packages by biocViews}
\usage{
get_packages_by_view(view, pkg_list = NULL)
}
\arguments{
\item{view}{One biocView term.}

\item{pkg_list}{Value of a call to \code{biocPkgList()}.
If \code{NULL} (default), will call \code{biocPkgList()} internally.}
}
\value{
Character vector of package names.
}
\description{
Get Packages by biocViews
}
\examples{
get_packages_by_view("Spatial")
}
