% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/wordcloud.R
\name{get_view_cooccurrences}
\alias{get_view_cooccurrences}
\title{Get biocViews Co-Occurrence Counts}
\usage{
get_view_cooccurrences(view, pkg_list = NULL, keep_self = FALSE, ratio = FALSE)
}
\arguments{
\item{view}{A biocViews term.}

\item{pkg_list}{Value of a call to \code{biocPkgList()}.
If \code{NULL} (default), will call \code{biocPkgList()} internally.
See Details.}

\item{keep_self}{Logical. Include \code{view} itself in the output.}

\item{ratio}{Logical. Return fraction of packages annotated with \code{view} also
annotated with each other term, rather than raw counts.}
}
\value{
A tibble of two columns: \code{date} and \code{cooccurences}.
\code{package} is a biocViews term.
\code{value} is the number (or fraction, if \code{ratio = TRUE}) of packages associated with both \code{view} and \code{package}.
}
\description{
Get biocViews Co-Occurrence Counts
}
\details{
Calling \code{BiocPkgTools::biocPkgList()} and passing the result to
\code{get_packages_by_view()} or \code{get_packages_by_views()} is more efficient
if you are making multiple calls.
See vignette 'Optimisations' for a more comprehensive discussion and demonstration.
}
\examples{
library(dplyr)
get_view_cooccurrences("Spatial") |>
  arrange(desc(value))
}
