% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/get_views_counts_over_time.R
\name{get_view_counts_over_time}
\alias{get_view_counts_over_time}
\alias{get_views_counts_over_time}
\title{biocViews Counts Over Time}
\usage{
get_view_counts_over_time(view, pkg_list, pkg_years)

get_views_counts_over_time(views, pkg_list, pkg_years)
}
\arguments{
\item{view}{A single biocView term.}

\item{pkg_list}{Value of a call to \code{biocPkgList()}.
If \code{NULL} (default), will call \code{biocPkgList()} internally.
See Details.}

\item{pkg_years}{Value of a call to \code{getPkgYearsInBioc()}.
If \code{NULL} (default), will call \code{getPkgYearsInBioc()} internally.}

\item{views}{A character vector of biocView terms.}
}
\value{
For \code{get_view_counts_over_time()}, a tibble of two columns: \code{date} and \code{count}.
\code{date} is a sequence of dates spaced by six months from 2006 to the current year.
\code{count} is the number of packages associated with \code{view} (or one of its child terms).

For \code{get_views_counts_over_time()}, a tibble of $N+1$ columns:
one column \code{date} and one column for each view in \code{views}.
}
\description{
biocViews Counts Over Time
}
\details{
Calling \code{BiocPkgTools::biocPkgList()} and passing the result to
\code{get_packages_by_view()} or \code{get_packages_by_views()} is more efficient
if you are making multiple calls.
See vignette 'Optimisations' for a more comprehensive discussion and demonstration.
}
\examples{
suppressPackageStartupMessages(library(BiocPkgTools))

bioc_pkg_list <- get_all_biocpkglist(verbose = FALSE)
bioc_years <- BiocPkgTools::getPkgYearsInBioc()

get_view_counts_over_time(
  view = "SingleCell",
  pkg_list = bioc_pkg_list,
  pkg_years = bioc_years
)
get_views_counts_over_time(
  views = c("Spatial", "SingleCell"),
  pkg_list = bioc_pkg_list,
  pkg_years = bioc_years
)
}
