% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/years_in_bioc.R
\name{get_view_counts_over_time}
\alias{get_view_counts_over_time}
\title{biocViews Counts Over Time}
\usage{
get_view_counts_over_time(view, pkg_list = NULL, pkg_years = NULL)
}
\arguments{
\item{view}{A single biocView term.}

\item{pkg_list}{Value of a call to \code{biocPkgList()}.
If \code{NULL} (default), will call \code{biocPkgList()} internally.
See Details.}
}
\value{
A named integer vector of counts of packages associated with \code{view} (or one of its child terms) over time,
named by date.
}
\description{
biocViews Counts Over Time
}
\details{
Calling \code{BiocPkgTools::biocPkgList()} and passing the result to
\code{get_packages_by_view()} or \code{get_packages_by_views()} is more efficient
if you are making multiple calls.
See vignette 'Optimisations' for a more comprehensive discussion and demonstration.
}
\examples{
view_over_time("SingleCell")
}
