% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/matrix.R
\name{get_view_membership_matrix}
\alias{get_view_membership_matrix}
\title{Package-by-biocView Membership Matrix}
\usage{
get_view_membership_matrix(pkg_list = NULL)
}
\arguments{
\item{pkg_list}{Value of a call to \code{biocPkgList()}.
If \code{NULL} (default), will call \code{biocPkgList()} internally.
See Details.}
}
\value{
A logical matrix indicating packages associated with each biocViews term (or one of its child terms).
}
\description{
Package-by-biocView Membership Matrix
}
\details{
Calling \code{BiocPkgTools::biocPkgList()} and passing the result to
\code{get_packages_by_view()} or \code{get_packages_by_views()} is more efficient
if you are making multiple calls.
See vignette 'Optimisations' for a more comprehensive discussion and demonstration.
}
\examples{
out <- biocview_by_package_matrix()
out[1:5, 1:5]
}
